+++ src/main/java/com/mong/project/controller/member/MemberController.java
@@ -1,10 +1,13 @@
 package com.mong.project.controller.member;
 
+import com.mong.project.domain.board.Board;
 import com.mong.project.domain.member.Member;
 import com.mong.project.dto.request.member.ChangePasswordRequest;
 import com.mong.project.dto.request.member.LoginRequest;
 import com.mong.project.dto.request.member.MemberJoinRequest;
+import com.mong.project.dto.response.board.GetBoardResponse;
 import com.mong.project.dto.response.member.LoginResponse;
+import com.mong.project.service.board.BoardService;
 import com.mong.project.service.member.MemberService;
 
 import lombok.extern.slf4j.Slf4j;
@@ -14,24 +17,29 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.ArrayList;
+import java.util.List;
+
 @Slf4j
 @RestController
 public class MemberController {
 
-    private MemberService memberService;
+    private final MemberService memberService;
+    private final BoardService boardService;
 
     @Autowired
-    public MemberController(MemberService memberService){
+    public MemberController(MemberService memberService, BoardService boardService){
         this.memberService = memberService;
+        this.boardService = boardService;
     }
 
-    @GetMapping("/member")
+    @GetMapping("/members")
     public ResponseEntity<Member> getMember(@RequestParam(name = "member_id") Long memberId) {
         Member member = memberService.getMemberById(memberId);
         return ResponseEntity.ok().body(member);
     }
 
-    @PostMapping("/member")
+    @PostMapping("/members")
     public ResponseEntity<Void> join(@RequestBody MemberJoinRequest memberJoinRequest) {
         memberService.join(memberJoinRequest.toMember());
 
@@ -46,21 +54,33 @@ public class MemberController {
         return ResponseEntity.ok().body(loginResponse);
     }
 
-    @PutMapping("/member/password")
+    @PutMapping("/members/password")
     public ResponseEntity<Void> changePassword(@RequestBody ChangePasswordRequest changePasswordRequest) {
         memberService.changePasswd(changePasswordRequest.getMemberId(), changePasswordRequest.getNewPasswd());
 
         return ResponseEntity.ok().build();
     }
 
-    @PostMapping(value = "/member/image", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<Void> changeImage(@RequestParam(name = "member_id") Long id, @RequestParam(name = "image") MultipartFile image) {
-        memberService.setImage(id, image);
+    @GetMapping("/members/board")
+    public ResponseEntity<List<GetBoardResponse>> getMemberBoards(@RequestParam(name = "member_id") Long memberId) {
+        List<Board> boards = boardService.getBoardsByMemberId(memberId);
+        List<GetBoardResponse> responses = new ArrayList<>();
+        boards.forEach(board -> {
+            GetBoardResponse response = new GetBoardResponse(board);
+            responses.add(response);
+        });
+
+        return ResponseEntity.ok().body(responses);
+    }
+
+    @PostMapping(value = "/members/image", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    public ResponseEntity<Void> changeImage(@RequestParam(name = "member_id") Long memberId, @RequestParam(name = "image") MultipartFile image) {
+        memberService.setImage(memberId, image);
 
         return ResponseEntity.ok().build();
     }
 
-    @DeleteMapping("/member/image")
+    @DeleteMapping("/members/image")
     public ResponseEntity<Void> deleteImage(@RequestParam(name = "member_id") Long memberId) {
         memberService.deleteImage(memberId);
